Refactored ROV Control
